rename private_tools
