+++ bluemira/geometry/_private_tools.py
